Finished Sabotage
Subtract a number X from every element in the input array.
Then, use DP to find the maximum subarray sum for the new subtracted array.
Using that number, you can find the sum of all elements not in the max subarray.
If sum is positive, then the elements not in subarray have an avg > X
If sum is negative, then the elements have an avg < X
If the sum is 0, then the elements have an avg = X
Thus, we can binary search on X in log time
and check whether the avg is <, >, or = x in linear time